1.紧急修复0.0.9版本中的空指针异常。 2.完善说明文档。

+++ hasor-rsf/src/main/java/net/hasor/rsf/protocol/DataField.java
@@ -42,7 +42,7 @@ public class DataField implements ProtocolCode {
     //
     //
     public void decode(ByteBuf buf) throws Throwable {
-        int length = buf.readBytes(2).readInt();
+        int length = buf.readBytes(2).readShort();
         if (length != 0) {
             this.value = buf.readBytes(length).toString(RSFConstants.DEFAULT_CHARSET);
         }
